lab2/requirements.txt
altgraph==0.17.2
async-generator==1.10
attrs==21.4.0
blinker==1.4
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.2
future==0.18.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
kaitaistruct==0.9
outcome==1.1.0
pbkdf2==1.3
pefile==2022.5.30
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.14.1
pycryptodome-test-vectors==1.0.8
pycryptodomex==3.14.1
pydivert==2.1.0
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.6
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PySide6==6.3.0
PySide6-Addons==6.3.0
PySide6-Essentials==6.3.0
PySocks==1.7.1
pywin32-ctypes==0.2.0
requests==2.27.1
rsa==4.8
selenium==4.1.5
selenium-wire==4.6.4
shiboken6==6.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
tk==0.1.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.5.4
websocket-client==1.3.2
Werkzeug==2.0.3
wsproto==1.1.0
zstandard==0.17.0
